requirements.txt
torch
torchvision
timm
numpy
opencv-python
scikit-learn
matplotlib
tqdm
giotto-tda
pyyaml